refactor(map_viewModel): move map related view models to map folder
