+++ test/mocks/overrides/override_pin_view_model.dart
@@ -1,6 +1,6 @@
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/models/ui/map_pin_details.dart";
-import "package:proxima/viewmodels/map_pin_view_model.dart";
+import "package:proxima/viewmodels/map/map_pin_view_model.dart";
 
 import "../data/map_pin.dart";
